requirements.txt
torch==2.1.2
tensorboard
wheel
python-dotenv
transformers==4.38.2
datasets==2.16.1
accelerate==0.26.1
evaluate==0.4.1
bitsandbytes==0.42.0
trl==0.7.11
peft==0.8.2

